// types/meta.ts
export namespace Facebook {
  /**
   * Props that are going to be passed to facebook's tracking function
   */
  export interface Event<T extends Event.EventName> {
    /**
     * The name of the event.
     */
    event_name: T
    /**
     * The custom data associated with the event.
     */
    custom_data?: Event.CustomData<T>
    /**
     * The ID of the event.
     */
    event_id?: string
    /**
     * @default The current timestamp in seconds.
     */
    event_time?: number
    /**
     * @default The current URL.
     */
    event_source_url?: string
    user_data?: Event.UserData
    opt_out?: boolean
    /**
     * @default "website"
     */
    action_source?: string
    data_processing_options?: string[]
    data_processing_options_country?: number
    data_processing_options_state?: number
    advanced_measurement_table?: string
    advertiser_tracking_enabled?: boolean
    messaging_channel?: string
  }

  export type BrowserEvent<T extends Event.EventName> = Omit<
    Event<T>,
    | 'event_time'
    | 'event_source_url'
    | 'user_data'
    | 'opt_out'
    | 'action_source'
    | 'data_processing_options'
    | 'data_processing_options_country'
    | 'data_processing_options_state'
    | 'advanced_measurement_table'
    | 'advertiser_tracking_enabled'
    | 'messaging_channel'
  >

  export namespace Event {
// ...
    export type EventName =
      | 'AddPaymentInfo'
      | 'AddToCart'
      | 'AddToWishlist'
      | 'CompleteRegistration'
      | 'Contact'
      | 'CustomizeProduct'
      | 'Donate'
      | 'FindLocation'
      | 'InitiateCheckout'
      | 'Lead'
      | 'PageView'
      | 'Purchase'
      | 'Schedule'
      | 'Search'
      | 'StartTrial'
      | 'SubmitApplication'
      | 'Subscribe'
      | 'ViewContent'
      | (string & {})

    /**
     * Represents the event data for a specific event name.
     * @template T - The event name.
     * @param T - The event name.
     * @returns The event data for the specified event name.
     */
    export type CustomData<T extends EventName> = T extends 'AddPaymentInfo'
      ? AddPaymentInfo
      : T extends 'AddToCart'
        ? AddToCart
        : T extends 'AddToWishlist'
          ? AddToWishlist
          : T extends 'CompleteRegistration'
            ? CompleteRegistration
            : T extends 'InitiateCheckout'
              ? InitiateCheckout
              : T extends 'Lead'
                ? Lead
                : T extends 'Purchase'
                  ? Purchase
                  : T extends 'Search'
                    ? Search
                    : T extends 'StartTrial'
                      ? StartTrial
                      : T extends 'Subscribe'
                        ? Subscribe
                        : T extends 'ViewContent'
                          ? ViewContent
                          : T extends 'PageView'
                            ? PageView
                            : T extends 'Contact'
                              ? Contact
                              : T extends 'CustomizeProduct'
                                ? CustomizeProduct
                                : T extends 'Donate'
                                  ? Donate
                                  : T extends 'FindLocation'
                                    ? FindLocation
                                    : T extends 'Schedule'
                                      ? Schedule
                                      : T extends 'SubmitApplication'
                                        ? SubmitApplication
                                        : FacebookData

    export interface UserData {
      email?: string
      phone?: string
      gender?: string
      first_name?: string
      last_name?: string
      date_of_birth?: string
      /**
       * The user's city.
       * @default Retrieved from request headers using `x-vercel-ip-city`, `cf-ipcity`, or `x-city`.
       */
      city?: string
      /**
       * The user's state or province.
       * @default Retrieved from request headers using `x-vercel-ip-country-region`, `cf-region`, or `x-region`.
       */
      state?: string
      /**
       * The user's zip or postal code.
       */
      zip?: string
      /**
       * The user's country.
       * @default Retrieved from request headers using `x-vercel-ip-country`, `cf-ipcountry`, or `x-country`.
       */
      country?: string
      external_id?: string
      /**
       * @default Try to get the IP address from the request headers using the
       * `x-forwarded-for` or the `x-real-ip` header.
       */
      client_ip_address?: string
      /**
       * @default Try to get the user agent from the request headers using the
       * `user-agent` header.
       */
      client_user_agent?: string
      /**
       * @default Try to get the `_fbp` cookie value.
       */
      fbp?: string
      /**
       * @default Try to get the `_fbc` cookie value.
       */
      fbc?: string
      subscription_id?: string
      fb_login_id?: string
      lead_id?: string
      dobd?: string
      dobm?: string
      doby?: string
      madid?: string
      anon_id?: string
      app_user_id?: string
      ctwa_clid?: string
      page_id?: string
    }
// ...
  }
}